refactor(backend): group rust modules by domain
src-tauri/src/ regrouped into audio/, library/, persist/ submodules:
- audio/{formats,player}.rs (was audio_formats.rs + player.rs)
- library/{db,scanner,scan_state}.rs
- persist/{config,session}.rs
- playlist.rs stays flat (no peers)

Internal refs use super:: for siblings, crate::<group>::<mod> across groups.
lib.rs mod block + use statements updated. CLAUDE.md architecture section refreshed.
