+++ src-tauri/src/audio/mod.rs
@@ -1,0 +1,2 @@
+pub mod formats;
+pub mod player;
