+++ src-tauri/src/library/mod.rs
@@ -1,0 +1,3 @@
+pub mod db;
+pub mod scan_state;
+pub mod scanner;
