+++ src-tauri/src/persist/mod.rs
@@ -1,0 +1,2 @@
+pub mod config;
+pub mod session;
