+++ src-tauri/src/persist/session.rs
