+++ src-tauri/src/playlist.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use std::collections::HashSet;
 
-use crate::db::{Db, Track};
+use crate::library::db::{Db, Track};
 
 const JINGLE_EVERY: i64 = 4;
 const COMMERCIAL_EVERY: i64 = 8;
